Create board.h with the board struct
Create a header file containing the main board struct. A chessboard
in this engine will be represented using bitboards. Bitboards are
generally faster than an array of 64 pieces because of bitwise
operators. Most modern computers are 64-bit, which make bitwise
operations on 64-bit numbers very fast. The board struct currently
has only a few variables in it, but more will be added later.