Add setBoardToFen to board.c
Add a function that allows the user to set up a chessboard to the
position held by a FEN string. This function can be used to set the
board to any position provided it is a legal chess position.

Currently, some data in the FEN is ignored (castle permissions, en
passant square, and the fifty move count). Castling, en passant
captures, draw by 50 moves, and draw by repetition will all be
added later so I can focus on writing and testing the more basic
rules first.