Add a Makefile
Add a Makefile to the project for easy compilation. I am using gcc
version 7.4.0 on WSL. The makefile can create a debug version and a
release version. In the debug version, NDEBUG is not defined and
debug.c is included in the compilation. In the release version,
NDEBUG is defined and debug.c is not included in the compilation.
For both versions, all warnings are enabled by using -Wall and
-Wextra as compilation flags.

-Wpedantic is used to stick to the C standard, for portability. In
the future, compiler-specific builtin functions and attributes will
be used to increase code speed and efficiency.