Add macros for the move flags
There are 5 different move flags: (1) The capture flag is a 1 if
the move was a capture move. (2) The promotion flag is a 1 if the
move was a pawn promotion (a pawn reached the opponent's side of
the board). (3) The castle flag is a 1 if the move was a castle
move. (4) The en passant flag is a 1 if the move was an en passant
capture. (5) The pawn start flag is a 1 if the move was a pawn
moving 2 spaces forward from its starting square. The only case
where 2 flags can be set at the same time is a simultaneous capture
and promotion. All other moves will have either 1 or 0 flags set.