Create attack.c
Create attack.c with the getKingAttacks function. getKingAttacks
uses the kingAttacks array (an array of bitboards such that the
i-th bitboard has a 1 bit in all positions that a king at index i
would be able to attack) and the getLSB function (a function that
returns the index of the least significant 1 bit in a bitboard).

All of the attack functions (getKnightAttacks, getPawnAttacks, etc)
will use similar arrays to getKingAttacks, but a slightly different
method is needed for sliding pieces (rook, bishop, queen) whose
attacks depend on the position of other pieces on the board, not
just their own position.