Add knight attacks to attack.c
Create the function getKnightAttacks() in attack.c. This function
uses the knightAttacks array to find the squares that a knight on
a given position is attacking. getKnightAttacks() does not store
information to indicate where an attack comes from. This means
that the function is most useful when used with only 1 knight,
since the caller knows where the knight is and where it attacks.