Change parameter for getKnightAttacks()
Change the getKnightAttacks() function in attacks.c to only take in
the position of a single knight instead of a bitboard of multiple
knights. The caller will have to call getKnightAttacks() multiple
times, but now they can know which squares are attacked by which
knight. Also update the function declaration in attacks.h.