Update the pawn attack functions
Split the getPawnAttacksRight() and getPawnAttacksLeft() into
separate functions for each color. Now there are 4 pawn attack
functions (2 for each color and 2 for each direction). In
movegne.c, there will be separate functions to generate the white
and black pawn attacks, so it makes sense that the pawn attack
functions are also separated by color.