Begin implementing move generation.
In movegen.c, implement move generation for kings and knights. Most
of the code for move generation will be in the generateAllMoves()
function. This function is designed to remove if statements and
branching based on which color's turn it is (besides one if
statement at the start of the function).