Add squareAttacked() to attack.c
Add a function to check if a square (or multiple squares) are being
attacked by a certain side. This function is useful for checking if
moves are legal. For example, if white makes a move that causes the
square of the white king to be under attack by any of the black
pieces, that move is illegal. This function will be used by the
makeMove() function to check if a move is legal or not.