Add a history member to the board struct
Add an array to the board struct that allows it to store moves that
are made. These moves will be used by the undoMove() function
whenever an illegal move is made by makeMove() and if the user ever
wants to take back a move. It will also be used to check for
repetitions.