Add perft testing
Add a file perft.c that can be used to test the move generator.
perft.c has its own main function and will only be used for testing
purposes. The file creates a board, reads in a bunch of FEN
strings, and generates moves up to a certain depth on the board. It
then counts how many moves it generated and compares this value to
a precomputed answer. The move generator for this project is bug
free when all the perft tests pass.