Add a pieces array to defs.c
Add an array to defs.c that can allow the user to quickly get the
integer (enum) value of a piece if they know its color and type.
This is useful for generating promotion moves, since we want to
generate a move for each of the piece types (except pawn and king)
of a certain color. Also add an extern definition in defs.h so that
the array can be used anywhere in the project.