Implement pawn move generation
In movegen.c, add functions to generate moves for pawns. These
functions will be separate from the other pieces because there are
a lot of special rules regarding pawns, such as en passant, pawn
starts (moving 2 squares forward if it is the pawn's first move),
and sideways attacks.

The board needs to keep track of an en passant square if the move
is legal, so add a member to the board struct. We will also need
to store the en passant square in the previousMove struct so that
it can be retrieved if a move is undone. The makeMove() and
undoMove() functions still need to be updated to handle pawn moves.