Add castle permissions to the board struct
Add a member variable to the board struct (and the PreviousMove
struct) that will keep track of which castling moves are legal for
the current position. Only the lowest 4 bits of the int will be
used. These bits correspond to bit flags in defs.h. If the bit
in the castlePerms int is 1, then that castling move is legal.