Update setBoardToFen for castlePerms
Update the setBoardToFen() function to scan in the castle
permissions and the en passant square of the FEN string. Before,
these parts of the FEN string were ignored. Now, the castle
permissions and the en passant square of the board can be set by
calling this function. Also, update the debug functions to check
the board's castlePerms member variable.