Add GCC-specific builtin functions
Most C compilers come with builtin functions that only work for
that compiler. GCC comes with some very helpful functions that
allow me to very quickly find the index of the least significant
bit, the most significant bit, and the number of bits set to 1
in an integer. Replacing the code in getLSB(), getMSB(), and
countBits() with these builtin functions makes move generation for
this engine 4 times faster. getMSB() and countBits() do not have
much of an impact, but getLSB() is called multiple times for every
move that is generated.