Change the size of a move to 32 bits
Decrease the size of a move integer from 64 bits to 32 bits.
Originally, 39 bits were allocated for the move score, but that
many is unnecessary. There are not very many types of moves
in chess (less than 100) so using the remaining 7 bits of a 32 bit
integer will suffice. Change the moveList to contain moves of only
32 bits and update every function that uses moves.