Add multithreading to perft.c
Allow perft tests to be run on multiple threads by using gcc's
pthread library. For each test position, the number of threads
created will be equal to the number of legal moves possible in
that position. This may be inefficient, since some tests are
creating around 50 threads. In the future I might limit the number
of threads to be 4 or 8 (or some changeable number) since most
modern processors have around 4-8 cores.

Also, there is a bug with timer.c where the calculated time to
complete a test is way higher than it should be. This may be caused
by each thread adding to the timer at the same time. Ex: If there
are 5 threads running, the time increases 5 times faster than
normal.