Create initializeAll() in defs.c
Create a function to call all the necessary initialization
functions for this chess engine. Currently there are functions to
initialize the bishop and rook attack tables and to initialize the
hash keys in hash.c. In defs.c create a function initializeAll() to
call all these functions, so that only one function needs to be
called by the user at the start of the program.