Modify timing for multithreaded perft tests.
Instead of timing the perft multithreaded tests with the cpu's
clock cycles, use the current calendar time in seconds. This is
only accurate to the second and timing the tests this way include
printing the results and setting up the boards. The displayed time
to run all the tests will be a little longer than it should be and
rounded down to the second, but will still be much more accurate
than before.

With this new more accurate timing method, we can see that
multithreading makes the perft tests run 3 times faster than using
a single thread. This can probably be sped up even further by using
a hash table.