Add getTime() to defs.c
Add a timer function to defs.c. This function will be used to time
how long it takes the alpha-beta algorithm to search to a certain
depth. Knowing this time allows the program to know when to stop
searching for a better move based on a search time limit. The timer
will also be used to time the perft tests. The timer function will
use an os-specific timing function to retrieve the time.