Update paramater for generatePositionKey
Change the paramater for generatePositionKey() from a const Board*
to a const void*. There was a conflict where hash.h was including
board.h and board.h was including hash.h. Change this function in
hash.h to take a void* instead of a Board* so that it does not
have to include board.h.