Add moveExists() to board.c
Add a function in board.c to check if a given move is legal in a
given position. This function will be used to check that the moves
stored in the hash tables are legal. In the event of a zobrist key
clash for 2 different boards, this function will prevent errors by
double checking that the move is still legal in the position.