Create a new file evaluate.c
Evaluate.c will contain everything that is used to evaluate a
chess position. These evaluations are used by the alphaBeta
algorithm to determine which move to make in the current position.
These evaluations are important because a wrong evaluation of a
position could lead to a blunder in that or previous positions.